Variable time zone
Instead of assuming the timestamp is based on GMT-6, it now searches for timestamp based on a variable time zone, the time zone in effect when the reservation was made. So, for example, it searches based on GMT-6 for reservations made before daylight savings time ended on Nov 3, 2024 and GMT-7 for reservations made after Nov 3.